requirements.txt
Flask==3.0.3
pandas==2.2.2
openpyxl==3.1.5
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httplib2==0.22.0
packaging==24.1
gunicorn==22.0.0
XlsxWriter==3.2.0
